Refactor Fileset Complete
